Refactor Version
