+++ cmd/spectre/cmd.go
@@ -1,4 +1,4 @@
-//  Copyright (C) 2020 Maker Ecosystem Growth Holdings, INC.
+//  Copyright (C) 2021-2023 Chronicle Labs, Inc.
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU Affero General Public License as
@@ -18,7 +18,7 @@ package main
 import (
 	"github.com/spf13/cobra"
 
-	suite "github.com/chronicleprotocol/oracle-suite"
+	"github.com/chronicleprotocol/oracle-suite/cmd"
 	"github.com/chronicleprotocol/oracle-suite/pkg/config/spectre"
 	"github.com/chronicleprotocol/oracle-suite/pkg/log/logrus/flag"
 )
@@ -32,7 +32,7 @@ type options struct {
 func NewRootCommand(opts *options) *cobra.Command {
 	rootCmd := &cobra.Command{
 		Use:           "spectre",
-		Version:       suite.Version,
+		Version:       cmd.Version,
 		Short:         "",
 		Long:          ``,
 		SilenceErrors: false,
