+++ cmd/spire/main.go
@@ -1,4 +1,4 @@
-//  Copyright (C) 2020 Maker Ecosystem Growth Holdings, INC.
+//  Copyright (C) 2021-2023 Chronicle Labs, Inc.
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU Affero General Public License as
@@ -18,11 +18,11 @@ package main
 import (
 	"os"
 
-	suite "github.com/chronicleprotocol/oracle-suite"
+	"github.com/chronicleprotocol/oracle-suite/cmd"
 )
 
 func main() {
-	opts := options{Version: suite.Version}
+	opts := options{Version: cmd.Version}
 	rootCmd := NewRootCommand(&opts)
 
 	if err := rootCmd.Execute(); err != nil {
