--- shell.nix
@@ -1,2 +1,0 @@
-{ pkgs ? import <nixpkgs> { }, oracle-suite ? pkgs.callPackage ./default.nix { } }:
-pkgs.mkShell { buildInputs = [ pkgs.jq oracle-suite ]; }
